feat: migrate files for rest_api/test_tasks
